chore: better organize telemetry utilities
